Fix insufficient dtype in HDF export
As it turns out, GPS times such as "1000000042" cannot accurately
be stored as float32 ("f4") numbers, which means that the
"event_time" will get corrupted when saving the generated data as
an HDF file. This commit fixes the data type for event_times to
float64, which should solve the problem.
Additionally, all injection_parameters are now also stored as
float64 numbers. Technically, this *shouldn't* be necessary, since
these parameters should all live in a more reasonable value range,
but better safe than sorry :)